feat: refactored the peers table to be in line with the new design
